Create README.md
Euclidean-Distance-Metric
This program was designed as a subtask of an ongoing project which is the development of the unsupervised K - Means Clustering Algorithm using Python.
This sub task (Euclidean Distance Project) enables python enthusiasts to get into the internal working of lists and insight into other data collection types.
A key takeaway from this project is that when CONVERTING lists into dictionaries, repeated keys will be removed automatically. This posed an issue moving forward with the mathematical calculations in the specific sequence required in the concluding sections of the program. The solution was initiating a for loop and zipping lists together whilst iterating throughout the elements within the lists in the sequence required required for specific calculations (see line 80) | Python Jupyter Notebook file. 
Furthermore developed is the same program using Java Script which proves very useful for comparing programming languages on many different levels.